refactor(sync): extract sync/pczt.rs
Move the hardware-wallet PCZT pipeline out of sync/mod.rs into a
dedicated pczt submodule:

  - create_pczt_from_proposal  (consume proposal → base PCZT)
  - discard_proposal           (idempotent cleanup for cancel paths)
  - add_proofs_to_pczt         (Orchard + optional Sapling proofs)
  - redact_pczt_for_signer     (strip witnesses / proprietary meta
                                 before handing to Keystone)
  - extract_and_broadcast_pczt (combine + broadcast + persist)

The module docstring restates the four pipeline invariants that were
documented in CLAUDE.md and in the previous mod.rs comment block, so
anyone touching this file sees the broadcast-before-persist ordering,
the post-broadcast fallback storage path, the Sapling-params
dual-wiring rule, and the consume-on-entry PROPOSAL_STORE contract
without having to go hunt for them.

Visibility is `pub(crate)` — api/sync.rs still calls into these by
name. `PROPOSAL_STORE`, `StoredProposal`, `open_wallet_db`, and
`WalletDatabase` stay in sync/mod.rs as `pub(super)` so both
sync/send.rs and sync/pczt.rs can reach them without a
cross-submodule cycle.

The PROPOSAL_STORE lifecycle tests (discard-idempotent, consume-on-
entry, replay rejection) stay in sync/mod.rs because they exercise
the shared store directly. They call into the re-exported functions
via `super::*`.

sync/mod.rs is down from 868 to 484 lines. 23 tests pass (including
the 3 PROPOSAL_STORE regression tests, which now call into pczt.rs
through the re-exports).
