+++ rust/src/wallet/sync/mod.rs
@@ -1,5 +1,3 @@
-use std::convert::Infallible;
-
 use rand::rngs::OsRng;
 use zcash_client_backend::{
     data_api::{
@@ -18,13 +16,17 @@ use zcash_client_sqlite::{
     util::SystemClock,
 };
 use zcash_primitives::block::BlockHash;
-use zcash_proofs::prover::LocalTxProver;
 use zcash_protocol::consensus::{BlockHeight, Network};
 
 use crate::wallet::keys::parse_account_uuid;
 
+mod pczt;
 mod send;
 
+pub(crate) use pczt::{
+    add_proofs_to_pczt, create_pczt_from_proposal, discard_proposal, extract_and_broadcast_pczt,
+    redact_pczt_for_signer,
+};
 pub(crate) use send::{estimate_fee, execute_proposal, propose_send};
 
 pub(super) type WalletDatabase = WalletDb<rusqlite::Connection, Network, SystemClock, OsRng>;
@@ -247,309 +249,6 @@ pub(super) static PROPOSAL_STORE: std::sync::LazyLock<Mutex<ProposalStore>> =
 pub(super) struct ProposalStore {
     pub proposals: HashMap<u64, StoredProposal>,
     pub next_id: u64,
-}
-
-// ======================== PCZT (Hardware Wallet) ========================
-
-/// Create a PCZT from a stored proposal (for hardware wallet signing).
-///
-/// This is the hardware-wallet analogue of `execute_proposal`, and mirrors
-/// its lifecycle: the proposal is **removed** from the store on entry, so
-/// any subsequent failure (PCZT creation error, hardware signing cancel,
-/// broadcast rejection) can't leave a replayable proposal ID behind. If the
-/// caller aborts the send flow before reaching this function (e.g. the
-/// confirmation dialog is cancelled), Dart is expected to call
-/// `discard_proposal` explicitly to release the stored proposal.
-pub fn create_pczt_from_proposal(
-    db_path: &str, network: Network, proposal_id: u64,
-) -> Result<Vec<u8>, String> {
-    use zcash_client_backend::data_api::wallet::create_pczt_from_proposal as zcb_create_pczt;
-    use zcash_client_backend::wallet::OvkPolicy;
-
-    // Consume the proposal up-front (matches execute_proposal), so that any
-    // later failure path leaves the PROPOSAL_STORE clean.
-    let stored = {
-        let mut store = PROPOSAL_STORE.lock().map_err(|e| format!("Lock: {e}"))?;
-        store.proposals.remove(&proposal_id)
-            .ok_or("Proposal not found (expired or already consumed)")?
-        // lock dropped here, before heavy work
-    };
-
-    let mut db = open_wallet_db(db_path, network)?;
-    let pczt = zcb_create_pczt::<_, _, Infallible, _, Infallible, _>(
-        &mut db, &network, stored.account_id, OvkPolicy::Sender, &stored.proposal,
-    ).map_err(|e| format!("Create PCZT failed: {e}"))?;
-
-    Ok(pczt.serialize())
-}
-
-/// Release a stored proposal without executing it. Called from the Dart
-/// send flow when the user cancels before `create_pczt_from_proposal`
-/// (e.g. dismisses the confirmation dialog, cancels the Sapling params
-/// download prompt). Idempotent: safe to call for a proposal that has
-/// already been consumed or never existed.
-pub fn discard_proposal(proposal_id: u64) {
-    if let Ok(mut store) = PROPOSAL_STORE.lock() {
-        store.proposals.remove(&proposal_id);
-    }
-}
-
-/// Add Orchard (and, if needed, Sapling) proofs to a PCZT locally.
-/// Returns a PCZT-with-proofs, which must later be combined with the
-/// signed PCZT returned by the hardware signer.
-///
-/// Sapling params paths are only required when the PCZT contains a non-empty
-/// Sapling bundle (e.g. the recipient is a Sapling-only address or a Unified
-/// Address without an Orchard receiver). Orchard-only sends can pass `None`
-/// for both paths. This matches the Zashi / zcash-android-wallet-sdk
-/// hardware-wallet flow: the hardware device only signs Orchard spends, the
-/// phone generates all ZK proofs.
-pub fn add_proofs_to_pczt(
-    pczt_bytes: &[u8],
-    spend_params_path: Option<&str>,
-    output_params_path: Option<&str>,
-) -> Result<Vec<u8>, String> {
-    use pczt::roles::prover::Prover;
-
-    let pczt = pczt::Pczt::parse(pczt_bytes)
-        .map_err(|e| format!("Parse PCZT: {e:?}"))?;
-
-    let mut prover = Prover::new(pczt);
-
-    if prover.requires_orchard_proof() {
-        prover = prover
-            .create_orchard_proof(&orchard::circuit::ProvingKey::build())
-            .map_err(|e| format!("Orchard proof: {e:?}"))?;
-    }
-
-    if prover.requires_sapling_proofs() {
-        match (spend_params_path, output_params_path) {
-            (Some(sp), Some(op)) if !sp.is_empty() && !op.is_empty() => {
-                let local_prover = LocalTxProver::new(
-                    std::path::Path::new(sp),
-                    std::path::Path::new(op),
-                );
-                prover = prover
-                    .create_sapling_proofs(&local_prover, &local_prover)
-                    .map_err(|e| format!("Sapling proofs: {e:?}"))?;
-            }
-            _ => {
-                return Err(
-                    "PCZT requires Sapling proofs but no Sapling params were supplied. \
-                     Download sapling-spend.params and sapling-output.params first."
-                        .into(),
-                );
-            }
-        }
-    }
-
-    Ok(prover.finish().serialize())
-}
-
-/// Redact information from a PCZT that the signer role doesn't need
-/// (witnesses, proprietary metadata). Produces the bytes to send to
-/// the hardware wallet for signing.
-pub fn redact_pczt_for_signer(pczt_bytes: &[u8]) -> Result<Vec<u8>, String> {
-    use pczt::roles::redactor::Redactor;
-
-    let pczt = pczt::Pczt::parse(pczt_bytes)
-        .map_err(|e| format!("Parse PCZT: {e:?}"))?;
-
-    let redacted = Redactor::new(pczt)
-        .redact_global_with(|mut r| r.redact_proprietary("zcash_client_backend:proposal_info"))
-        .redact_orchard_with(|mut r| {
-            r.redact_actions(|mut ar| {
-                ar.clear_spend_witness();
-                ar.redact_output_proprietary("zcash_client_backend:output_info");
-            });
-        })
-        .redact_sapling_with(|mut r| {
-            r.redact_spends(|mut sr| sr.clear_witness());
-            r.redact_outputs(|mut or| {
-                or.redact_proprietary("zcash_client_backend:output_info");
-            });
-        })
-        .redact_transparent_with(|mut r| {
-            r.redact_outputs(|mut or| {
-                or.redact_proprietary("zcash_client_backend:output_info");
-            });
-        })
-        .finish();
-
-    Ok(redacted.serialize())
-}
-
-/// Combine a PCZT-with-proofs and a PCZT-with-signatures, broadcast the
-/// resulting transaction, and persist it to the wallet DB **only after** the
-/// broadcast is accepted.
-///
-/// Ordering is critical here. The naive "store then broadcast" path leaves
-/// the wallet in an inconsistent state when lightwalletd rejects the tx:
-/// the DB thinks the notes are spent, the network never saw the tx, and
-/// the user has to manually recover. Instead we:
-///   1. Parse + combine + extract the final `Transaction` (validates proofs)
-///      using only in-memory state.
-///   2. Broadcast the raw bytes. If this fails, return Err — DB untouched.
-///   3. After broadcast is accepted, persist to the wallet DB. The primary
-///      path is `extract_and_store_transaction_from_pczt`, which preserves
-///      rich PCZT recipient/memo metadata (the "sent to …" display info).
-///      If it fails — e.g. because the PCZT proprietary metadata got
-///      rewritten by the hardware signer, or the DB layer is temporarily
-///      unhappy — fall back to `decrypt_and_store_transaction`, which is
-///      the same path sync uses when it discovers one of our sent txs on
-///      the chain. That preserves correctness (spent notes are marked
-///      spent) at the cost of losing the rich metadata. Only if *both*
-///      storage paths fail do we surface an error to the user, because at
-///      that point the wallet genuinely cannot tell the user the truth
-///      about their balance.
-///
-/// Sapling params paths are required whenever the combined PCZT contains a
-/// non-empty Sapling bundle — e.g. a send to a Sapling-only recipient, or a
-/// Unified Address without an Orchard receiver. In that case both the
-/// in-memory `TransactionExtractor` and the on-disk
-/// `extract_and_store_transaction_from_pczt` need Sapling verifying keys;
-/// we load them once from `LocalTxProver` and reuse the references. Orchard-
-/// only sends can pass `None` for both paths and the Sapling code path is
-/// skipped entirely. This matches the `add_proofs_to_pczt` contract: if the
-/// caller supplied params there, they must supply them here too.
-pub async fn extract_and_broadcast_pczt(
-    db_path: &str, lightwalletd_url: &str, network: Network,
-    pczt_with_proofs_bytes: &[u8],
-    pczt_with_signatures_bytes: &[u8],
-    spend_params_path: Option<&str>,
-    output_params_path: Option<&str>,
-) -> Result<String, String> {
-    use pczt::roles::combiner::Combiner;
-    use pczt::roles::tx_extractor::TransactionExtractor;
-    use zcash_client_backend::data_api::wallet::{
-        decrypt_and_store_transaction, extract_and_store_transaction_from_pczt,
-    };
-
-    // Re-parsing and re-combining is cheap compared to ZK proof validation;
-    // we do it twice so we can hand one owned Pczt to the in-memory extractor
-    // and a second owned Pczt to the storage function after broadcast.
-    fn combine_pczts(proofs: &[u8], sigs: &[u8]) -> Result<pczt::Pczt, String> {
-        let p = pczt::Pczt::parse(proofs)
-            .map_err(|e| format!("Parse PCZT with proofs: {e:?}"))?;
-        let s = pczt::Pczt::parse(sigs)
-            .map_err(|e| format!("Parse PCZT with signatures: {e:?}"))?;
-        Combiner::new(vec![p, s])
-            .combine()
-            .map_err(|e| format!("Combine PCZTs: {e:?}"))
-    }
-
-    let orchard_vk = orchard::circuit::VerifyingKey::build();
-
-    // Load Sapling verifying keys once if the caller supplied params. The
-    // prover keeps the underlying params alive, and `verifying_keys()`
-    // returns owned `(SpendVerifyingKey, OutputVerifyingKey)`. We'll hand
-    // references into this tuple to both `TransactionExtractor::with_sapling`
-    // and `extract_and_store_transaction_from_pczt`.
-    let sapling_vks: Option<(
-        sapling_crypto::circuit::SpendVerifyingKey,
-        sapling_crypto::circuit::OutputVerifyingKey,
-    )> = match (spend_params_path, output_params_path) {
-        (Some(sp), Some(op)) if !sp.is_empty() && !op.is_empty() => {
-            let prover = LocalTxProver::new(
-                std::path::Path::new(sp),
-                std::path::Path::new(op),
-            );
-            Some(prover.verifying_keys())
-        }
-        _ => None,
-    };
-
-    // Step 1: extract the Transaction without touching the DB. We keep `tx`
-    // around after broadcast so the fallback storage path can use it.
-    let tx = {
-        let mut extractor = TransactionExtractor::new(
-            combine_pczts(pczt_with_proofs_bytes, pczt_with_signatures_bytes)?,
-        )
-        .with_orchard(&orchard_vk);
-        if let Some((spend_vk, output_vk)) = sapling_vks.as_ref() {
-            extractor = extractor.with_sapling(spend_vk, output_vk);
-        }
-        extractor
-            .extract()
-            .map_err(|e| format!("Extract TX from PCZT: {e:?}"))?
-    };
-    let txid = tx.txid();
-
-    let tx_bytes = {
-        let mut buf = Vec::new();
-        tx.write(&mut buf).map_err(|e| format!("Serialize TX: {e}"))?;
-        buf
-    };
-
-    // Step 2: broadcast. On any failure here, the DB is untouched, so the
-    // wallet's view of spendable notes is unchanged.
-    let (mut client, _tor_guard) =
-        crate::wallet::sync_engine::open_lwd_channel(lightwalletd_url)
-            .await
-            .map_err(|e| e.to_string())?;
-
-    let resp = client.send_transaction(
-        zcash_client_backend::proto::service::RawTransaction {
-            data: tx_bytes,
-            height: 0,
-        }
-    ).await.map_err(|e| format!("Broadcast: {e}"))?.into_inner();
-
-    // zebra-lightwalletd returns the txid in `error_message` on success, so the
-    // only reliable signal is `error_code`.
-    if resp.error_code != 0 {
-        return Err(format!(
-            "Broadcast rejected: {} (code {})",
-            resp.error_message, resp.error_code
-        ));
-    }
-
-    // Step 3: broadcast was accepted. Persist locally so the UI sees the
-    // tx immediately and the spent notes stop showing up as spendable.
-    let mut db = open_wallet_db(db_path, network)?;
-
-    // Primary path: rich PCZT-aware storage (preserves recipient/memo).
-    // Hand Sapling verifying keys in whenever the combined PCZT has a
-    // Sapling bundle, otherwise librustzcash rejects the extraction with
-    // `SaplingRequired` before we can store anything.
-    let sapling_vk_pair = sapling_vks.as_ref().map(|(s, o)| (s, o));
-    match extract_and_store_transaction_from_pczt::<_, zcash_client_sqlite::ReceivedNoteId>(
-        &mut db,
-        combine_pczts(pczt_with_proofs_bytes, pczt_with_signatures_bytes)?,
-        sapling_vk_pair,
-        Some(&orchard_vk),
-    ) {
-        Ok(_) => return Ok(txid.to_string()),
-        Err(primary_err) => {
-            log::warn!(
-                "keystone: PCZT-aware storage failed after broadcast \
-                 (txid={txid}): {primary_err}. Falling back to chain-style \
-                 decrypt_and_store_transaction; rich recipient metadata \
-                 will not be available in history until the next sync."
-            );
-
-            // Fallback path: same code sync uses when it discovers a
-            // wallet tx on the chain. Marks spent notes correctly via
-            // nullifier matching and picks up any change note back to
-            // us from enc_ciphertext decryption. The recipient/memo
-            // metadata that was only in the PCZT proprietary fields is
-            // lost, but correctness is preserved — the spent notes no
-            // longer appear spendable.
-            decrypt_and_store_transaction(&network, &mut db, &tx, None)
-                .map_err(|fallback_err| {
-                    format!(
-                        "Broadcast succeeded (txid={txid}) but both local \
-                         storage paths failed. Primary: {primary_err}. \
-                         Fallback: {fallback_err}. The transaction is on \
-                         the network; check an explorer to confirm, and \
-                         do not attempt to send again until the next \
-                         sync reconciles your balance."
-                    )
-                })?;
-        }
-    }
-
-    Ok(txid.to_string())
 }
 
 // ======================== Diversified Address ========================
